feat：切出分支 v1.0。

+++ src/views/personal-notice/index.vue
@@ -1,82 +1,83 @@
 <template>
-  <c-page-label title="系统公告" icon="cyber-gonggao" document-link="javascript:;">
-    <template #tips>维护公告相关信息。</template>
-  </c-page-label>
+  <div>
+    <c-page-label title="系统公告" icon="cyber-gonggao" document-link="#消息中心-开发中">
+      <template #tips>维护公告相关信息。</template>
+    </c-page-label>
 
-  <c-radio-wrapper v-model:value="queryState.type">
-    <c-radio-button value="">全部</c-radio-button>
-    <c-radio-button value="1" dark>系统公告</c-radio-button>
-    <c-radio-button value="2" dark>系统公告</c-radio-button>
-    <c-radio-button value="3" dark>待办事项</c-radio-button>
-  </c-radio-wrapper>
+    <c-radio-wrapper v-model:value="queryState.type">
+      <c-radio-button value="">全部</c-radio-button>
+      <c-radio-button value="1" dark>系统公告</c-radio-button>
+      <c-radio-button value="2" dark>系统公告</c-radio-button>
+      <c-radio-button value="3" dark>待办事项</c-radio-button>
+    </c-radio-wrapper>
 
-  <div class="sys-notice-card">
-    <div class="sys-notice-list">
-      <div class="sys-notice-list-header">
-        <a-checkbox
-          class="mr-16px"
-          v-model:checked="noticeState.checkAll"
-          :disabled="!noticeState.dataSource?.length"
-          :indeterminate="!noticeState.checkAll && !!noticeState.value?.length"
-          @change="methods.checkAllHandler"
-        ></a-checkbox>
-        <a-button :disabled="!noticeState.value?.length" @click="methods.batchDelete()">删除</a-button>
-        <a-button :disabled="!noticeState.value?.length">标记已读</a-button>
-        <a-button :disabled="!noticeState.dataSource?.length">全部已读</a-button>
-      </div>
-      <div class="sys-notice-list-container">
-        <a-checkbox-group v-model:value="noticeState.value" class="w-1/1" @change="methods.changeCheckbox">
-          <a-empty v-if="!noticeState.dataSource?.length"></a-empty>
-          <div
-            v-else
-            v-for="(item, index) in noticeState.dataSource"
-            :key="index"
-            :class="[
-              'sys-notice-item',
-              { 'sys-notice-item-unpublish': item.status == '1' },
-              { 'sys-notice-item-publish': item.status == '0' },
-              { 'sys-notice-item-select': queryState.id == item.id },
-            ]"
-            @click="methods.onSelect(item)"
-          >
-            <a-checkbox :value="item.id"></a-checkbox>
-            <c-icon icon="cyber-zhengchang" size="8" class="mr-10px ml-10px"></c-icon>
-            <div class="w-0 flex-1">
-              <div class="w-1/1 flex-between">
-                <span class="ellipsis">{{ item.noticeTitle || '-' }}</span>
-                <span class="notice-status">[<c-cell-dict :options="$dictStore.noticeStatus" :value="item.status" style="min-height: initial;" />]</span>
+    <div class="sys-notice-card">
+      <div class="sys-notice-list">
+        <div class="sys-notice-list-header">
+          <a-checkbox
+            class="mr-16px"
+            v-model:checked="noticeState.checkAll"
+            :disabled="!noticeState.dataSource?.length"
+            :indeterminate="!noticeState.checkAll && !!noticeState.value?.length"
+            @change="methods.checkAllHandler"
+          ></a-checkbox>
+          <a-button :disabled="!noticeState.value?.length" @click="methods.batchDelete()">删除</a-button>
+          <a-button :disabled="!noticeState.value?.length">标记已读</a-button>
+          <a-button :disabled="!noticeState.dataSource?.length">全部已读</a-button>
+        </div>
+        <div class="sys-notice-list-container">
+          <a-checkbox-group v-model:value="noticeState.value" class="w-1/1" @change="methods.changeCheckbox">
+            <a-empty v-if="!noticeState.dataSource?.length"></a-empty>
+            <div
+              v-else
+              v-for="(item, index) in noticeState.dataSource"
+              :key="index"
+              :class="[
+                'sys-notice-item',
+                { 'sys-notice-item-unpublish': item.status == '1' },
+                { 'sys-notice-item-publish': item.status == '0' },
+                { 'sys-notice-item-select': queryState.id == item.id },
+              ]"
+              @click="methods.onSelect(item)"
+            >
+              <a-checkbox :value="item.id"></a-checkbox>
+              <c-icon icon="cyber-zhengchang" size="8" class="mr-10px ml-10px"></c-icon>
+              <div class="w-0 flex-1">
+                <div class="w-1/1 flex justify-between">
+                  <span class="ellipsis">{{ item.noticeTitle || '-' }}</span>
+                  <span class="notice-status">[<c-cell-dict :options="NOTICE_STATUS" :value="item.status" style="min-height: initial;" />]</span>
+                </div>
+                <div class="mt-7px text-[#63637299]">{{ item.createTime || '-' }}</div>
               </div>
-              <div class="mt-7px text-[#63637299]">{{ item.createTime || '-' }}</div>
             </div>
-          </div>
-        </a-checkbox-group>
+          </a-checkbox-group>
+        </div>
       </div>
-    </div>
-    <div class="sys-notice-body">
-      <template v-if="queryState.id">
-        <div class="sys-notice-body-header">
-          <div>
-            <div class="text-[#242E42] text-16px">{{ noticeState.detail.noticeTitle || '-' }}</div>
-            <div class="mt-8px text-[#A8ACB8]">
-              <span>发布范围：{{ noticeState.detail.noticeScope || '-' }}</span>
-              <span class="ml-30px">发布时间：{{ noticeState.detail.status == '0' ? noticeState.detail.createTime : '-' }}</span>
+      <div class="sys-notice-body">
+        <template v-if="queryState.id">
+          <div class="sys-notice-body-header">
+            <div>
+              <div class="text-[#242E42] text-16px">{{ noticeState.detail.noticeTitle || '-' }}</div>
+              <div class="mt-8px text-[#A8ACB8]">
+                <span>发布范围：{{ noticeState.detail.noticeScope || '-' }}</span>
+                <span class="ml-30px">发布时间：{{ noticeState.detail.status == '0' ? noticeState.detail.createTime : '-' }}</span>
+              </div>
             </div>
           </div>
-        </div>
-        <div class="sys-notice-body-container">
-          <span>{{ noticeState.detail.noticeContent }}</span>
-        </div>
-      </template>
+          <div class="sys-notice-body-container">
+            <span>{{ noticeState.detail.noticeContent }}</span>
+          </div>
+        </template>
+      </div>
     </div>
   </div>
 </template>
 
 <script setup>
-import { dictStore } from '@/store';
 import { message } from 'ant-design-vue';
-import { changeHistoryState, initHistoryState } from 'cyber-web-ui';
-import axios, { batchOperateConfrim, deleteConfrim, queryDetail } from '@/api';
-const $dictStore = dictStore();
+import { changeHistoryState, initHistoryState, useDict, Modal } from 'cyber-web-ui';
+import axios, { queryDetail } from '@/api';
+const { NOTICE_STATUS } = useDict({ COMMON: ['NOTICE_STATUS'] });
 const modifyRef = ref(); // 弹窗ref
 // 表格请求参数
 const queryState = reactive({
@@ -130,15 +131,16 @@ const methods = {
   },
   // 删除
   delete(record) {
-    deleteConfrim({
+    Modal.verify({
       content: `是否确认删除“${record.creator}（${record.noticeTitle}）”的公告？`,
       value: record.noticeTitle,
-    }, {
-      url: '/auth/sysnotice',
-      method: 'delete',
       params: {
-        id: record.id,
-      }
+        url: '/auth/sysnotice',
+        method: 'delete',
+        params: {
+          id: record.id,
+        }
+      },
     }).then(() => {
       queryState.id = undefined;
       noticeState.detail = {};
@@ -148,30 +150,41 @@ const methods = {
   },
   // 批量删除
   batchDelete() {
-    batchOperateConfrim('确定要批量删除选中的公告吗，删除后将无法恢复！', {
-      url: '/auth/sysnotice/batch',
-      method: 'delete',
-      data: {
-        ids: noticeState.value,
-      }
-    }).then(() => {
-      noticeState.value = [];
-      noticeState.checkAll = false;
-      methods.searchQuery();
+    Modal.confirm({
+      content: '确定要批量删除选中的公告吗，删除后将无法恢复！',
+      onOk: async () => {
+        try {
+          axios.request({
+            url: '/auth/sysnotice/batch',
+            method: 'delete',
+            data: {
+              ids: noticeState.value,
+            }
+          });
+          message.success(res.message);
+          noticeState.value = [];
+          noticeState.checkAll = false;
+          methods.searchQuery();
+          return Promise.resolve();
+        } catch {
+          return Promise.reject();
+        }
+      },
     });
   },
   // 切换状态
   changeStatus(record, status, text) {
-    deleteConfrim({
+    Modal.verify({
       content: `是否确认${text}“${record.creator}（${record.noticeTitle}）”的公告？`,
       value: record.noticeTitle,
-    }, {
-      url: '/auth/sysnotice/batch',
-      method: 'put',
-      data: {
-        ids: [record.id],
-        status: status
-      }
+      params: {
+        url: '/auth/sysnotice/batch',
+        method: 'put',
+        data: {
+          ids: [record.id],
+          status: status
+        }
+      },
     }).then(() => {
       methods.searchQuery();
       methods.queryNoticeDetail(record.id);
@@ -190,17 +203,27 @@ const methods = {
       message.warning(`没有可${status == '0' ? '发布' : '撤回'}的公告！`);
       return;
     }
-    batchOperateConfrim(text, {
-      url: '/auth/sysnotice/batch',
-      method: 'put',
-      data: {
-        ids: ids,
-        status: status,
-      }
-    }).then(() => {
-      noticeState.value = [];
-      noticeState.checkAll = false;
-      methods.searchQuery();
+    Modal.confirm({
+      content: text,
+      onOk: async () => {
+        try {
+          axios.request({
+            url: '/auth/sysnotice/batch',
+            method: 'put',
+            data: {
+              ids: ids,
+              status: status,
+            }
+          });
+          message.success(res.message);
+          noticeState.value = [];
+          noticeState.checkAll = false;
+          methods.searchQuery();
+          return Promise.resolve();
+        } catch {
+          return Promise.reject();
+        }
+      },
     });
   },
 };
@@ -251,24 +274,24 @@ const methods = {
           bottom: 0;
           left: 16px;
         }
-        .cyber-icon {
+        .cyber-iconfont {
           transition: all 0.2s linear;
         }
         &.sys-notice-item-publish {
           .notice-status,
-          .cyber-icon {
+          .cyber-iconfont {
             color: @success-color;
           }
         }
         &.sys-notice-item-unpublish {
           .notice-status,
-          .cyber-icon {
+          .cyber-iconfont {
             color: @error-color;
           }
         }
         &.sys-notice-item-select {
           background: #E4E8EF;
-          .cyber-icon {
+          .cyber-iconfont {
             color: transparent;
           }
         }
